protect classes api

+++ app/api/classes/[id]/route.ts
@@ -6,38 +6,42 @@ type ParamsProps = {
   };
 };
 export const DELETE = async (req: NextRequest, { params }: ParamsProps) => {
-  try {
-    const user = await prisma.class.findUnique({
-      where: {
-        id: params.id,
-      },
-    });
-    if (!user) {
-      return NextResponse.json(
-        { message: "Class with this id doesn't exist" },
-        { status: 404 }
-      );
-    }
-
+  if (req.headers.get("apiKey")) {
     try {
-      await prisma.class.delete({
+      const user = await prisma.class.findUnique({
         where: {
           id: params.id,
         },
       });
-      return NextResponse.json({
-        message: "Your desired class successfully deleted!",
-      });
+      if (!user) {
+        return NextResponse.json(
+          { message: "Class with this id doesn't exist" },
+          { status: 404 }
+        );
+      }
+
+      try {
+        await prisma.class.delete({
+          where: {
+            id: params.id,
+          },
+        });
+        return NextResponse.json({
+          message: "Your desired class successfully deleted!",
+        });
+      } catch (error) {
+        return NextResponse.json(
+          { message: "There is an error in server" },
+          { status: 500 }
+        );
+      }
     } catch (error) {
       return NextResponse.json(
         { message: "There is an error in server" },
         { status: 500 }
       );
     }
-  } catch (error) {
-    return NextResponse.json(
-      { message: "There is an error in server" },
-      { status: 500 }
-    );
+  } else {
+    return NextResponse.json({ message: "Access denied", status: 401 });
   }
 };
